requirements.txt
Flask
pandas
scikit-learn
requests
python-dotenv
google-genai